Rework input to use a channel
This changes line-input to use a "channel", a self-contained read/write stream where the interpretter does not need to manage the details of where the characters written / read go.

This should allow for interpretting from strings in RAM and from the disk, without extra code paths and without having to copy the whole string first.

Some RAM definitions were reorganised as they were not C64 or CBM-specific.

Binary size: 2614b